User-Movies Database Relationship

+++ src/main/java/com/obss/movieTracker/repository/MovieRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MovieRepository extends CrudRepository<Movie, Long> {
+public interface MovieRepository extends CrudRepository<Movie, Integer> {
     List<Movie> findByMovieName(String movieName);
 }
